fix: GetViewportColorTexID uint32→uint64 — SIGSEGV on Metal
ImGui::Image received a truncated 32-bit texture pointer instead of the
full 64-bit id<MTLTexture>. The upper 32 bits were lost in the uint32_t
return type, producing an invalid pointer → SIGSEGV on first ImGui draw.

RenderFacade::GetViewportColorTexID() now returns uint64_t.
EditorViewport uses reinterpret_cast<ImTextureID> for the correct cast.

+++ CHEngine/src/CHEngine/Render/RenderFacade.cpp
@@ -313,7 +313,7 @@ namespace CHEngine
         return s_ViewportOutputTex;
     }
 
-    uint32_t RenderFacade::GetViewportColorTexID()
+    uint64_t RenderFacade::GetViewportColorTexID()
     {
         if (!s_ViewportOutputTex.IsValid()) return 0;
         IRenderFactory* f = GetFactory();
